skills/aicrate/git-vcs-wiki/git-vcs-wiki/scripts/requirements.txt
PyGithub==2.9.1
python-gitlab==8.3.0
